+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]## [0.5.4.6b2]
 ## Added
 - dse_do_utils.core with Core01DataManager, Core01OptimizationEngine, Core01EnvironmentManager
-### Changed
+### Fixed
+- Fix ScenarioDbManager.duplicate_scenario_in_db future warning
 
 ## [0.5.4.6b1]- 2023-02-06
 ## Added