+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.5.4.6b2]
+## [Unreleased]## [0.5.4.6b3]
+
+## [0.5.4.6b2]- 2023-02-13
 ## Added
 - dse_do_utils.core with Core01DataManager, Core01OptimizationEngine, Core01EnvironmentManager
 ### Fixed